split files frm to custom post into mulitple files, added all sanitization needed, made functions to for part of code that can be reused (getting question info, pritning to test file), got rid of necessary code and comments
