+++ form_to_custom_post.php
@@ -2,7 +2,7 @@
 require 'table_utils.php';
 require 'table_writing_functions.php';
 require 'table_reading_functions.php';
-require 'community_reviews_custom_post.php';
+require 'insert_custom_post.php';
 
 function profile_info_sub( $record, $ajax_handler ){
     global $wpdb;
@@ -48,29 +48,20 @@ function fluent_summit_review_from_sub($entryId, $formData, $form) {
     $file = fopen($file_path, "w") or die('fopen failed');
     //fwrite($file,"fluent_summit_review_from_sub\n\n");
     $existing_form_names = get_all_form_names($file);
-    //$form_name_str = print_r($existing_form_names, true);
-    //fwrite($file, "registered form names:\n".$form_name_str."\n");
+    
     $current_form_name = $form->title; 
-    //fwrite($file, $current_form_name."\n");
+    
     $current_form_id = $form->id;
 
-    //$form_p = print_r($form, true);
-    //$form_fields = $form->get( "form_fields" );
-    $formData_p = print_r($formData, true);
-    //fwrite($file, "formData 1: \n".$formData_p."\n\n");
-
     $qs_and_as = fluent_get_fields_array($formData, $file);
-    //$qs_and_as_p = print_r($qs_and_as, true);
-    //fwrite($file, "questions and answers: \n".$qs_and_as_p."\n\n");
     
     if(in_array($current_form_name, $existing_form_names)){
-        //fwrite($file, "Starting the Process \n\n");
-        //-----------------write to tables-----------------------------
+        //-----------------write to tables (table_writing_functions)-----------------------------
         $review_id = summit_form_submission_write_to_tables($current_form_id, $qs_and_as, $file);
-        //---------------create custom post-----------------------------------
+        //---------------create custom post (insert_custom_post)-----------------------------------
         $header = summit_form_submission_custom_post_content($review_id, $current_form_id, $qs_and_as, $file);
-        $header_info_read = print_r($header, true);
-        fwrite($file, "HEADER: \n".$header_info_read."\n\n");
+        //$header_info_read = print_r($header, true);
+        //fwrite($file, "HEADER: \n".$header_info_read."\n\n");
         insert_into_ski_review($header, $file, $current_form_name);
     }
     fclose($file);
@@ -97,21 +88,14 @@ function elementor_summit_review_from_sub( $record, $ajax_handler ) {
     $file_path = plugin_dir_path( __FILE__ ) . '/testfile.txt'; 
     $file = fopen($file_path, "w") or die('fopen failed');
     $existing_form_names = get_all_form_names($file);
-    //fwrite($file,"form names: ".$existing_form_names."\n");
 
     $current_form_name = $record->get_form_settings( 'form_name' ); 
-    //fwrite($file,"current form name: ".$current_form_name."\n");
 
     $current_form_id = $record->get_form_settings('form_id');
-    //fwrite($file,"current form id: ".$current_form_id."\n");
 
     $raw_fields = $record->get( 'fields' );
-    //$raw_fields_p = print_r($raw_fields, true);
-    //fwrite($file, "elementor raw fields: \n".$raw_fields_p."\n\n");
 
     $qs_and_as = elementor_get_fields_array($raw_fields, $file);
-    //$q_and_a_p = print_r($qs_and_as, true);
-    //fwrite($file, "elementor q and a: \n".$q_and_a_p."\n\n");
 
     if(in_array($current_form_name, $existing_form_names)){
     	//-----------------write to tables----------------------------
@@ -119,8 +103,6 @@ function elementor_summit_review_from_sub( $record, $ajax_handler ) {
 
         //---------------create custom post-----------------------------------
         $header = summit_form_submission_custom_post_content($review_id, $current_form_id, $qs_and_as, $file);
-        //$header_info_read = print_r($header, true);
-        //fwrite($file, "HEADER: \n".$header_info_read."\n\n");
     	insert_into_ski_review($header, $file, $current_form_name);
 	}
     fclose($file);
@@ -142,15 +124,11 @@ function get_all_form_names($file){
     //fwrite($file, $start);
     $review_forms_table = $wpdb->prefix . "bcr_review_forms";
     $q = $wpdb->prepare("SELECT reviewFormName FROM $review_forms_table;");
-    //$q = "SELECT reviewFormName FROM `wp_bcr_review_forms`";
-    //$wpdb->prepare("SELECT %s FROM %s;", array("reviewFormName", $review_table));
     $res = $wpdb->get_results($q);
-    $return_array = [];//json_decode(json_encode($res),true);
+    $return_array = [];
     foreach ($res as $value) {
         $return_array[] = $value->reviewFormName;
     }
-    //$res_p = print_r($return_array, true);
-    //fwrite($file, "all form names:\n".$res_p."\n");
     return $return_array;
 }
 //---------------------------------------
@@ -158,54 +136,25 @@ function summit_form_submission_write_to_tables($current_form_id, $record, $file
     //$start = "\n\n start summit_form_submission_write_to_tabless \n";
     //fwrite($file, $start);
     $answer_ids = insert_into_answer_table($record, $file);
-    //fwrite($file, "write to answer table ids:\n".$answer_ids."\n\n") or die('fwrite 1 failed');
-    //$res1 = implode(", ", $answer_ids)."\n";
-    //fwrite($file, $res1) or die('fwrite 2 failed');
-
-    //$ajax_handler->add_response_data( true, $output );
-    //fwrite($file, "write to review table id:\n") or die('fwrite 3 failed');
     $id = insert_into_review_table($current_form_id,$file);
-    //fwrite($file, "last inserted review id: ".strval($id)."\n") or die('fwrite 2 failed');
-    //$ajax_handler->add_response_data( true, $output );
     insert_into_review_answer_table($id, $answer_ids, $file);
-    //fwrite($file, "\n\nsummit_insert_into_review_answer_table function exited\n\n");
-    //$ajax_handler->add_response_data( true, $output );
     return $id;
 }
 
 function summit_form_submission_custom_post_content($current_review_id, $current_form_id,$record,$file){
     //$start = "\n\n summit_form_submission_custom_post_content \n";
     //fwrite($file, $start);
-    
     $product_info = [];
-    $product_info['productName'] = $record[2];//get_product_info($current_form_id,$file);
-    //$p_info_read = print_r($product_info, true);
-    //fwrite($file, "product info: \n".$p_info_read."\n\n");
-    //$p_info_read = print_r($product_info, true);
-    //fwrite($file, "product info: \n".$p_info_read."\n\n");
+    $product_info['productName'] = $record[2];
     
     $brand_name = $record[1];
     //$brand_info = get_brand_info($brand_name, $file);
-    //$b_info_read = print_r($brand_info, true);
-    //fwrite($file, "brand info: \n".$b_info_read."\n\n");
 
     $category_info = get_category_info($current_form_id, $file);
-    //$c_info_read = print_r($category_info, true);
-    //fwrite($file, "categroy info: \n".$c_info_read."\n\n");
-
-    /*
-    $sport_info = get_sport_info($category_info['sportID']);
-    //$s_info_read = print_r($sport_info, true);
-    //fwrite($file, "sport info: \n".$s_info_read."\n\n");
-    */
 
     $q_and_a_content = get_answer_and_question_content($record,$file);
-    //$qa_info_read = print_r($q_and_a_content, true);
-    //fwrite($file, "QandA info: \n".$qa_info_read."\n\n");
     
     $user_info = get_user_information($file);
-    //$u_info_read = print_r($user_info, true);
-    //fwrite($myfile, "user info: \n".$u_info_read."\n\n");
 
     $header = array(
         'reviewID' => $current_review_id,
