+++ table_writing_functions.php
@@ -11,9 +11,7 @@ function insert_into_answer_table($record,$file){
     //$start = "\n\n SUMMIT INSERT INTO ANSWER TABLE \n";
     //fwrite($file, $start);
     $answer_table = $wpdb->prefix . "bcr_answers";
-    $answer_ids = []; //used for when inserting into reviews answers
-    //$question_ids = [];//manually entered into elementor form, until we can make a form dynamically 
-    //$raw_fields = $record->get( 'fields' );
+    $answer_ids = [];
     $output = [];
     foreach($record as $id => $field ){
         $fields_answers = [];
@@ -21,11 +19,8 @@ function insert_into_answer_table($record,$file){
         $fields_answers['answerContent'] = $field;
         $output['success'] = $wpdb->insert($answer_table, $fields_answers);
         //if($output == success){
-
         //}
-        $last_answer_id = $wpdb->insert_id;//answer id generated upon entering into table
-        //!!!!!!may be susceptible to collisions with mulitple users
-        //maybe there is a lock for each form submitted idk
+        $last_answer_id = $wpdb->insert_id;
         array_push($answer_ids, $last_answer_id);
     }
     //$ajax_handler->add_response_data( true, $output );
@@ -44,27 +39,18 @@ function insert_into_review_table($RF_id, $file){
 
         $current_userID = get_current_userID($file);
         if(gettype($current_userID)=="string"){
-        //  fwrite($file,"user id grab failed: ".$current_userID."\n");
             die("user not found"); //should be a redirct to another page
         }
-        //fwrite($file,"userID = ".strval($current_userID)."\n");
         $fields_review['userID'] = $current_userID;
         
-        //$KTSid = get_knowthyself_id($current_userID);
-        //$fields_review['knowThyselfID'] = 4;//NOT NESSESARY ANYMORE get this id from bcr_know_thyself using userid HAS TO EXIST BEFORE SUBMISSION
-        //remove!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
         $fields_review['reviewFormID'] = $RF_id;
 
-        //fwrite($file, "fields_review: ".implode(", ",$fields_review)."\n");
+        
         $output2['success'] = $wpdb->insert($review_table, $fields_review);
-        //$str = "number of rows inserted: ".strval($output['success'])." ||| did not work if false\n";
-        //fwrite($file, $str);
-        //echo strval(output['success'])."<br>";
-        //$ajax_handler->add_response_data( true, $output );
-        $last_review_id = $wpdb->insert_id; //STILL NOT SURE IF THIS IS THREAD SAFE
+        $last_review_id = $wpdb->insert_id; 
         return $last_review_id;
 }
+
 /*
 function get_knowthyself_id($userID){
     global $wpdb;
@@ -74,6 +60,7 @@ function get_knowthyself_id($userID){
     return $res->knowThyselfID;
 }
 */
+
 function insert_into_review_answer_table($review_id, $answer_ids,$file){
     //insert answer ids int review answer table
         global $wpdb;
@@ -87,7 +74,6 @@ function insert_into_review_answer_table($review_id, $answer_ids,$file){
             $fields_review_answers['answerID'] = $id;
             $output['success'] = $wpdb->insert($review_answer_table, $fields_review_answers);
             //if($output == success){
-
             //}
         }
 }
